event: expose print_events and separate tests
