+++ crates/termoxide_event/src/crossterm.rs
@@ -5,7 +5,7 @@ use crossterm::{
     terminal::{disable_raw_mode, enable_raw_mode},
 };
 
-fn print_events(shutdown: &mpsc::Receiver<()>) -> io::Result<()> {
+pub fn print_events(shutdown: &mpsc::Receiver<()>) -> io::Result<()> {
     loop {
         if shutdown.try_recv().is_ok() {
             break;
@@ -30,28 +30,4 @@ pub fn read_events(shutdown: mpsc::Receiver<()>) -> io::Result<()> {
     }
 
     disable_raw_mode()
-}
-
-#[cfg(test)]
-mod tests {
-    use std::{sync::mpsc, thread, time::Duration};
-
-    use super::print_events;
-
-    #[test]
-    fn print_events_stops_when_shutdown_already_signaled() {
-        let (shutdown_tx, shutdown_rx) = mpsc::channel();
-        shutdown_tx.send(()).unwrap();
-
-        let (done_tx, done_rx) = mpsc::channel();
-        thread::spawn(move || {
-            let result = print_events(&shutdown_rx);
-            let _ = done_tx.send(result);
-        });
-
-        let result = done_rx.recv_timeout(Duration::from_secs(2)).expect(
-            "print_events did not return: the shutdown did not break the loop",
-        );
-        result.expect("print_events returned an error");
-    }
 }
